requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
colorama==0.4.6
colorlog==6.9.0
comtypes==1.4.11
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
einops==0.8.1
et-xmlfile==2.0.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
gputil==1.4.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
imagesize==1.4.1
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
lxml==6.0.0
markupsafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
multidict==6.6.4
mypy-extensions==1.1.0
networkx==3.5
numpy==2.2.6
openai==1.99.9
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.2
packaging==25.0
paddleocr==3.1.1
paddlepaddle==3.1.0
paddlex==3.1.4
paho-mqtt==2.1.0
pandas==2.3.1
pillow==11.3.0
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
protobuf==6.32.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pypdfium2==4.30.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyttsx3==2.99
pytz==2025.2
pywin32==311
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
sqlalchemy==2.0.43
srt==3.5.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
vosk==0.3.45
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
